remove pg version
