--- crates/github-data-graphql-database-pg/CHANGELOG.md
@@ -1,13 +1,0 @@
-# CHANGELOG
-
-# 0.2.1
-
-* rename env var to `PG_DATABASE_URL` 
-
-# 0.2.0
-
-* change log file table name to `osb_github_logfiles`
-
-# 0.1.0 
-
-* first version only pg
