--- crates/github-data-graphql-database-pg/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "github-data-graphql-database-pg"
-version = "0.2.1"
-edition = "2021"
-
-[dependencies]
-bson = "2.9.0"
-chrono = { version = "0.4", features = ["serde"] }
-clap = { version = "4.5.23", features = ["derive"] }
-diesel = { version = "2.2.6", features = ["postgres", "chrono", "uuid", "serde_json"] }
-diesel-async = { version = "0.5.2", features = ["postgres"] }
-dotenvy = "0.15.7"
-futures = "0.3"
-futures-core = "0.3.14"
-futures-executor = "0.3.14"
-futures-io = "0.3.21"
-futures-util = { version = "0.3.14", features = ["io"] }
-pq-sys = "0.6.1"
-rand = "0.8.4"
-reqwest = { version = "0.12.9", features = ["json"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_derive = "1.0.209"
-serde_json = "1.0"
-tokio = { version = "1", features = ["full"] }
-tokio-postgres = "0.7"
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-uuid = { version = "1.10.0", features = ["v4","fast-rng", "serde"] }
-
-[[bin]]
-name = "server"
-path = "src/main.rs"
