--- crates/github-data-graphql-database-pg/diesel.toml
@@ -1,9 +1,0 @@
-# For documentation on how to configure this file,
-# see https://diesel.rs/guides/configuring-diesel-cli
-
-[print_schema]
-file = "src/diesel_schema.rs"
-custom_type_derives = ["diesel::query_builder::QueryId", "Clone"]
-
-[migrations_directory]
-dir = "/workspaces/crates/github-data-graphql-database/migrations"
