--- crates/github-data-graphql-database-pg/sample.env
@@ -1,3 +1,0 @@
-PAT=ghp_
-PG_DATABASE_URL=postgresql://
-MONGO_DATABASE_URL=mongodb://
