--- crates/github-data-graphql-database-pg/src/diesel_schema.rs
@@ -1,19 +1,0 @@
-// @generated automatically by Diesel CLI.
-
-diesel::table! {
-    osb_github_logfiles (id) {
-        id -> Uuid,
-        logfile -> Jsonb,
-        created_at -> Timestamptz,
-        org_repo -> Text,
-    }
-}
-
-diesel::table! {
-    osb_user_custom_config (id) {
-        id -> Text,
-        user_id -> Text,
-        repo_name -> Text,
-        created_at -> Timestamptz
-    }
-}
