+++ crates/github-data-graphql-database/CHANGELOG.md
@@ -1,3 +1,9 @@
 # CHANGELOG
 
-* 0.1.0 - first version
+# 0.2.0
+
+* Remove mongo, keep everything in PG, use JSONB column type
+
+# 0.1.0
+
+* first version
