+++ crates/github-data-graphql-database/Cargo.toml
@@ -1,13 +1,13 @@
 [package]
 name = "github-data-graphql-database"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
 bson = "2.9.0"
 chrono = { version = "0.4", features = ["serde"] }
 clap = { version = "4.5.23", features = ["derive"] }
-diesel = { version = "2.2.6", features = ["postgres", "chrono", "uuid"] }
+diesel = { version = "2.2.6", features = ["serde_json","postgres", "chrono", "uuid"] }
 dotenv = "0.15.0"
 dotenvy = "0.15.7"
 futures = "0.3"
@@ -26,6 +26,7 @@ tokio = { version = "1", features = ["full"] }
 tokio-postgres = "0.7"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+uuid = { version = "1.11.0", features = ["serde"] }
 
 [[bin]]
 name = "server"
