+++ crates/github-data-graphql-database/src/diesel_schema.rs
@@ -6,3 +6,11 @@ diesel::table! {
         created_at -> Timestamptz
     }
 }
+diesel::table! {
+    osb_github_logfiles (id) {
+        id -> Uuid,
+        logfile -> Jsonb,
+        created_at -> Timestamptz,
+        org_repo -> Text,
+    }
+}
