+++ crates/github-data-graphql-database/src/pq_query.rs
@@ -1,8 +1,12 @@
 use diesel::pg::PgConnection;
+//use diesel::sql_types::Jsonb;
 use serde::{Serialize, Deserialize};
 use diesel::prelude::*;
 use crate::diesel_schema::osb_user_custom_config;
+use crate::diesel_schema::osb_github_logfiles;
 use diesel::result::Error;
+use serde_json::Value;
+
 #[derive(Debug, Serialize, Deserialize, Queryable, Selectable, QueryableByName)]
 #[diesel(table_name = osb_user_custom_config)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
@@ -11,6 +15,14 @@ pub struct Watch {
     pub user_id: String,
     pub repo_name: String,
     pub created_at: chrono::NaiveDateTime
+}
+
+#[derive(Insertable)]
+#[diesel(table_name = osb_github_logfiles)]
+pub struct NewLogfile{
+    #[diesel(sql_type = Jsonb)]
+    pub logfile: Value,
+    pub org_repo: String
 }
 
 pub fn get_connection(database_url: &str) -> PgConnection {
@@ -36,4 +48,16 @@ pub async fn get_repos(connection: &mut PgConnection, limit: i64, offset: i64) -
                 Err(e)
             }
         }
+}
+pub async fn insert_repos(connection: &mut PgConnection, repo: NewLogfile) -> Result<String, Error> {
+    use crate::diesel_schema::osb_github_logfiles::dsl::*;
+    let rows_inserted = diesel::insert_into(osb_github_logfiles)
+        .values(&repo)
+        .execute(connection)?;
+
+    if rows_inserted == 1 {
+        Ok(repo.org_repo)
+    } else {
+        Err(Error::NotFound)
+    }
 }
